Add extended User model and Video model
This adds several models:

First is the UserTable, which extends the built in User model with two new fields: role and points. Role is for storing the user's role as a string, and points is for storing the number of points the user has accrued

The Videos model has three fields: videoTitle, videoLink, videoPoints. These are for storing the title, link to the video and the number points given for watching respectively. There is also a videoThumbnailURL field to store the link to the video thumbnail (intended to be an image)

Another model, VideoWatchers, is used to identify which users have already watched what video. This way, we can make sure that points arent repeatedly awareded. This uses the id field from both the UserTable and Videos models as foreign keys, and adds a timestamp when the video was watched.