Rewrote QR Code View to account for Cooldown
Rewrote the QR Code view to add cooldown functionality:

This code works by first iterating though all records of the water station uses model - if it finds a record created within the cooldown period, it does not add the points.
If not, it adds the points as normal, and then adds a fill record to the model.

It also makes sure to erase obsolete records to prevent ballooning of the database.